requirements27.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
_license=1.1=py27_1
alabaster=0.7.9=py27_0
anaconda=4.3.1=np111py27_0
anaconda-client=1.6.0=py27_0
anaconda-navigator=1.5.0=py27_0
anaconda-project=0.4.1=py27_0
appnope=0.1.0=py27_0
appscript=1.0.1=py27_0
argcomplete=1.0.0=py27_1
astroid=1.4.9=py27_0
astropy=1.3=np111py27_0
babel=2.3.4=py27_0
backports=1.0=py27_0
backports_abc=0.5=py27_0
beautifulsoup4=4.5.3=py27_0
bitarray=0.8.1=py27_0
blaze=0.10.1=py27_0
bokeh=0.12.4=py27_0
boto=2.45.0=py27_0
bottleneck=1.2.0=np111py27_0
cdecimal=2.3=py27_2
cffi=1.9.1=py27_0
chardet=2.3.0=py27_0
chest=0.2.3=py27_0
click=6.7=py27_0
cloudpickle=0.2.2=py27_0
clyent=1.2.2=py27_0
colorama=0.3.7=py27_0
configobj=5.0.6=py27_0
configparser=3.5.0=py27_0
contextlib2=0.5.4=py27_0
cryptography=1.7.1=py27_0
curl=7.52.1=0
cycler=0.10.0=py27_0
cython=0.25.2=py27_0
cytoolz=0.8.2=py27_0
dask=0.13.0=py27_0
datashape=0.5.4=py27_0
decorator=4.0.11=py27_0
dill=0.2.5=py27_0
docutils=0.13.1=py27_0
entrypoints=0.2.2=py27_0
enum34=1.1.6=py27_0
et_xmlfile=1.0.1=py27_0
fastcache=1.0.2=py27_1
flask=0.12=py27_0
flask-cors=3.0.2=py27_0
freetype=2.5.5=2
funcsigs=1.0.2=py27_0
functools32=3.2.3.2=py27_0
futures=3.0.5=py27_0
get_terminal_size=1.0.0=py27_0
gevent=1.2.1=py27_0
greenlet=0.4.11=py27_0
grin=1.2.1=py27_3
h5py=2.6.0=np111py27_2
hdf5=1.8.17=1
heapdict=1.0.0=py27_1
icu=54.1=0
idna=2.2=py27_0
imagesize=0.7.1=py27_0
ipaddress=1.0.18=py27_0
ipykernel=4.5.2=py27_0
ipython=5.1.0=py27_1
ipython_genutils=0.1.0=py27_0
ipywidgets=5.2.2=py27_1
isort=4.2.5=py27_0
itsdangerous=0.24=py27_0
jbig=2.1=0
jdcal=1.3=py27_0
jedi=0.9.0=py27_1
jinja2=2.9.4=py27_0
jpeg=9b=0
jsonschema=2.5.1=py27_0
jupyter=1.0.0=py27_3
jupyter_client=4.4.0=py27_0
jupyter_console=5.0.0=py27_0
jupyter_core=4.2.1=py27_0
lazy-object-proxy=1.2.2=py27_0
libiconv=1.14=0
libpng=1.6.27=0
libtiff=4.0.6=3
libxml2=2.9.4=0
libxslt=1.1.29=0
llvmlite=0.15.0=py27_0
locket=0.2.0=py27_1
lxml=3.7.2=py27_0
markupsafe=0.23=py27_2
matplotlib=2.0.0=np111py27_0
mistune=0.7.3=py27_1
mkl=2017.0.1=0
mkl-service=1.1.2=py27_3
mpmath=0.19=py27_1
multipledispatch=0.4.9=py27_0
nbconvert=4.2.0=py27_0
nbformat=4.2.0=py27_0
networkx=1.11=py27_0
nltk=3.2.2=py27_0
nose=1.3.7=py27_1
notebook=4.3.1=py27_0
numba=0.30.1=np111py27_0
numexpr=2.6.1=np111py27_2
numpy=1.11.3=py27_0
numpydoc=0.6.0=py27_0
odo=0.5.0=py27_1
openpyxl=2.4.1=py27_0
openssl=1.0.2k=1
pandas=0.19.2=np111py27_1
partd=0.3.7=py27_0
path.py=10.0=py27_0
pathlib2=2.2.0=py27_0
patsy=0.4.1=py27_0
pep8=1.7.0=py27_0
pexpect=4.2.1=py27_0
pickleshare=0.7.4=py27_0
pillow=4.0.0=py27_0
pip=9.0.1=py27_1
ply=3.9=py27_0
prompt_toolkit=1.0.9=py27_0
psutil=5.0.1=py27_0
ptyprocess=0.5.1=py27_0
py=1.4.32=py27_0
pyasn1=0.1.9=py27_0
pyaudio=0.2.7=py27_0
pycosat=0.6.1=py27_1
pycparser=2.17=py27_0
pycrypto=2.6.1=py27_4
pycurl=7.43.0=py27_2
pyflakes=1.5.0=py27_0
pygments=2.1.3=py27_0
pylint=1.6.4=py27_1
pyopenssl=16.2.0=py27_0
pyparsing=2.1.4=py27_0
pyqt=5.6.0=py27_1
pytables=3.3.0=np111py27_0
pytest=3.0.5=py27_0
python=2.7.13=0
python-dateutil=2.6.0=py27_0
python.app=1.2=py27_4
pytz=2016.10=py27_0
pyyaml=3.12=py27_0
pyzmq=16.0.2=py27_0
qt=5.6.2=0
qtawesome=0.4.3=py27_0
qtconsole=4.2.1=py27_1
qtpy=1.2.1=py27_0
readline=6.2=2
redis=3.2.0=0
redis-py=2.10.5=py27_0
requests=2.12.4=py27_0
rope=0.9.4=py27_1
ruamel_yaml=0.11.14=py27_1
scandir=1.4=py27_0
scikit-image=0.12.3=np111py27_1
scikit-learn=0.18.1=np111py27_1
scipy=0.18.1=np111py27_1
seaborn=0.7.1=py27_0
setuptools=27.2.0=py27_0
simplegeneric=0.8.1=py27_1
singledispatch=3.4.0.3=py27_0
sip=4.18=py27_0
six=1.10.0=py27_0
snowballstemmer=1.2.1=py27_0
sockjs-tornado=1.0.3=py27_0
sphinx=1.5.1=py27_0
spyder=3.1.2=py27_0
sqlalchemy=1.1.5=py27_0
sqlite=3.13.0=0
ssl_match_hostname=3.4.0.2=py27_1
statsmodels=0.6.1=np111py27_1
subprocess32=3.2.7=py27_0
sympy=1.0=py27_0
terminado=0.6=py27_0
tk=8.5.18=0
toolz=0.8.2=py27_0
tornado=4.4.2=py27_0
traitlets=4.3.1=py27_0
unicodecsv=0.14.1=py27_0
wcwidth=0.1.7=py27_0
werkzeug=0.11.15=py27_0
wheel=0.29.0=py27_0
widgetsnbextension=1.2.6=py27_0
wrapt=1.10.8=py27_0
xlrd=1.0.0=py27_0
xlsxwriter=0.9.6=py27_0
xlwings=0.10.2=py27_0
xlwt=1.2.0=py27_0
xz=5.2.2=1
yaml=0.1.6=0
zlib=1.2.8=3
